refactor: Provide a sphinxnotes.any.api package (#43)
